+++ package.json
@@ -12,7 +12,8 @@
   "dependencies": {
     "axios": "^1.8.3",
     "react": "^18.2.0",
-    "react-dom": "^18.2.0"
+    "react-dom": "^18.2.0",
+    "react-router-dom": "^7.3.0"
   },
   "devDependencies": {
     "@types/react": "^18.0.37",
