+++ src/pages/newPost/NewPost.jsx
@@ -1,0 +1,7 @@
+function NewPost() {
+    return (
+        <div>form for new post</div>
+    );
+}
+
+export default NewPost;
