+++ src/pages/notFound/NotFound.jsx
@@ -1,0 +1,7 @@
+function NotFound() {
+    return (
+        <h1>Page not found</h1>
+    );
+}
+
+export default NotFound;
