+++ src/pages/post/Post.jsx
@@ -1,0 +1,12 @@
+import {useState} from "react";
+
+function Post() {
+    const [post, setPost] = useState({});
+    return (
+        <article>
+
+        </article>
+    );
+}
+
+export default Post;
